requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
boto3==1.20.2
botocore==1.23.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cryptography==35.0.0
dnspython==2.1.0
email-validator==1.1.3
et-xmlfile==1.1.0
fastapi==0.68.1
google-api-core==2.1.0
google-api-python-client==2.23.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
gunicorn==20.1.0
h11==0.12.0
httplib2==0.20.1
idna==3.2
iniconfig==1.1.1
jmespath==0.10.0
mysqlclient==2.0.3
numpy==1.21.2
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.2
pandas==1.3.3
pluggy==1.0.0
protobuf==3.18.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.2.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rope==0.20.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
uvicorn==0.15.0
